feat: #431 centralize write + log pre-process TU
