+++ src/main/scala/ca/cqdg/etl/processes/processes.scala
@@ -1,5 +1,8 @@
 package ca.cqdg.etl
 
+import bio.ferlab.datalake.spark3.config.Configuration
+import org.apache.spark.sql.{DataFrame, SaveMode}
+
 import java.util.concurrent.Executors
 import scala.concurrent.{ExecutionContext, ExecutionContextExecutorService}
 
@@ -13,4 +16,19 @@ package object processes {
 
   val tsv_with_headers = Map("sep" -> "\t", "header" -> "true")
   val json_multiline = Map("multiline" -> "true")
+
+  def write(sourceId: String, df: DataFrame, conf: Configuration): Unit = {
+    val source = conf.getDataset(sourceId)
+    val storage = conf.getStorage(source.storageid)
+    val outputPath = s"$storage/${source.path}"
+
+    df
+      .coalesce(1)
+      .write
+      .format(source.format.sparkFormat)
+      .mode(SaveMode.Overwrite)
+      .options(source.writeoptions)
+      .partitionBy(source.partitionby:_*)
+      .save(outputPath)
+  }
 }
